Implement array submodule
