+++ earthkit/meteo/geo/array/__init__.py
@@ -1,0 +1,1 @@
+from .geo import *  # noqa
