+++ earthkit/meteo/score/array/__init__.py
@@ -1,0 +1,1 @@
+from .crps import *  # noqa
