+++ earthkit/meteo/stats/array/__init__.py
@@ -1,0 +1,2 @@
+from .numpy_extended import *  # noqa
+from .quantiles import *  # noqa
