+++ earthkit/meteo/wind/array/__init__.py
@@ -1,0 +1,1 @@
+from .wind import *  # noqa
