+++ tests/geo/test_score.py
