+++ tests/solar/test_solar.py
