+++ tests/stats/test_stats.py
